fix(security): remove raw ADMIN_SECRET backdoor, call sync directly from cron
verifyAdmin accepted the raw ADMIN_SECRET as a permanent, unrevocable master
key that defeated JWT expiry/revocation. Removed it; admin API now accepts only
short-lived JWTs. The sync logic is extracted to lib/product-sync.ts so cron
calls runProductSync() directly instead of self-fetching /api/admin/scrape with
the secret. Admin login now compares credentials with a constant-time helper.
