+++ app/api/cron/sync/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 
 import { getConfig } from '@/lib/data-service';
+import { runProductSync } from '@/lib/product-sync';
 
 // Allow up to 60s on Vercel Pro; free plan caps at 10s but scrape is fire-and-forget
 export const maxDuration = 60;
@@ -29,23 +30,11 @@ export async function GET(req: NextRequest) {
     // If config check fails, proceed with sync anyway
   }
 
-  // Determine app URL for internal API call
-  const appUrl =
-    process.env.NEXT_PUBLIC_APP_URL ||
-    (process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : 'http://localhost:3000');
-
+  // Run the sync directly — no internal HTTP hop, no ADMIN_SECRET as a bearer token.
   try {
-    const res = await fetch(`${appUrl}/api/admin/scrape`, {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${process.env.ADMIN_SECRET}`,
-      },
-    });
-
-    const data = (await res.json()) as Record<string, unknown>;
-    console.log('[Cron] Auto-sync completed:', data);
-
-    return NextResponse.json({ success: res.ok, ...data });
+    const result = await runProductSync();
+    console.log('[Cron] Auto-sync completed:', result);
+    return NextResponse.json(result);
   } catch (error) {
     console.error('[Cron] Auto-sync failed:', error);
     return NextResponse.json({ error: 'Sync failed' }, { status: 500 });
